Remove heat endpoints from class group
Endpoints are managed for all other services via
coi::profiles::endpoints. Heat currently isn't
because its class structure doesn't match, which
is due to it having a separate class for the
cloudformation endpoint.

This patch should be paired with a patch to puppet-coi
and any other consumers of enabled_services to handle
this special case.

Change-Id: I9bbb1c59cbebc8962bdd34c890e0af1f1a4be469