Add ceilometer to stacktira
This patch adds the ceilometer_control
group to the stacktira controller.

The ceilometer::keystone::auth class is already
included in COI via enabled_services and is not
needed in stacktira due to regions_hash.

A singleton class is used to prevent multiple
central agents from running.

The keystone admin and public endpoints had to be
changed as the ceilometer client was adding
extra versioning to the path. This matches
the behavior in the comments and does not
seem to affect any other clients.

Change-Id: Ifd705c5efebcce8d80d40066139da284800d1fc7